# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS FinOps Analytics Platform API
  version: 1.0.0
  description: |
    The SaaS FinOps Analytics Platform provides multi-tenant cloud cost management,
    AI-powered database analytics, and subscription billing capabilities.

    ## Authentication

    The API supports two authentication methods:

    ### JWT Bearer Token
    Obtain a token via `POST /auth/login`. Include it in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Access tokens expire after **15 minutes**. Use `POST /auth/refresh` to obtain a new one.

    ### API Key
    Generate an API key via `POST /auth/api-keys`. Include it in the `X-API-Key` header:
    ```
    X-API-Key: <your_api_key>
    ```

    ## Rate Limiting

    Rate limits are enforced per user per minute based on subscription tier:

    | Plan       | Requests/min |
    |------------|-------------|
    | Free       | 100         |
    | Base       | 500         |
    | Pro        | 2,000       |
    | Enterprise | 10,000      |

    When the limit is exceeded, the API returns `429 Too Many Requests` with a `Retry-After` header.

  contact:
    name: Platform Support
    email: support@finops-platform.example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
  - url: http://localhost:8081
    description: Auth Service (direct)
  - url: http://localhost:8082
    description: Billing Service (direct)
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login. Expires in 15 minutes.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key generated via POST /auth/api-keys.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "invalid credentials"
        message:
          type: string
          example: "The provided email or password is incorrect"
      required: [error]
...
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          example: "rate limit exceeded"
        retry_after:
          type: integer
          example: 30
          description: Seconds until the rate limit resets
...
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stripe_invoice_id:
          type: string
          example: "in_stripe123"
        amount_cents:
          type: integer
          example: 2000
        currency:
          type: string
          example: "USD"
        status:
          type: string
          enum: [draft, open, paid, void, uncollectible]
          example: "paid"
        invoice_pdf_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
...
    CostSummaryResponse:
      type: object
      properties:
        total_cost:
          type: number
          format: float
          example: 15420.50
        currency:
          type: string
          example: "USD"
        start_date:
          type: string
          format: date
          example: "2024-01-01"
        end_date:
          type: string
          format: date
          example: "2024-01-31"
        breakdown_by_provider:
          type: object
          additionalProperties:
            type: number
          example:
            aws: 10200.00
            azure: 3120.50
            gcp: 2100.00
        breakdown_by_service:
          type: array
          items:
            type: object
            properties:
              service:
                type: string
              cost:
                type: number
          example:
            - service: "EC2"
              cost: 5000.00
            - service: "S3"
              cost: 1200.00
...
paths:
...
  /billing/invoices:
    get:
      tags: [billing]
      summary: List invoices
      description: Returns all invoices for the authenticated account, sorted by date descending.
      operationId: listInvoices
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Invoice list
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
                  total:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /billing/invoices/{id}/pdf:
    get:
      tags: [billing]
      summary: Download invoice PDF
      description: Redirects to the Stripe-hosted invoice PDF URL.
      operationId: downloadInvoicePdf
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Invoice ID
      responses:
        '302':
          description: Redirect to PDF URL
          headers:
            Location:
              schema:
                type: string
                format: uri
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - invoice belongs to another account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /finops/costs/summary:
    get:
      tags: [finops]
      summary: Get cost summary
      description: |
        Returns aggregated cloud cost data for the specified date range.
        Includes breakdown by provider and service.

        **Rate limit:** 100 req/min (Free), 500 (Base), 2000 (Pro), 10000 (Enterprise)
      operationId: getCostSummary
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2024-01-31"
        - name: provider
          in: query
          schema:
            type: string
            enum: [aws, azure, gcp]
          description: Filter by cloud provider
        - name: region
          in: query
          schema:
            type: string
          description: Filter by region (e.g., us-east-1)
          example: "us-east-1"
      responses:
        '200':
          description: Cost summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostSummaryResponse'
              example:
                total_cost: 15420.50
                currency: "USD"
                start_date: "2024-01-01"
                end_date: "2024-01-31"
                breakdown_by_provider:
                  aws: 10200.00
                  azure: 3120.50
                  gcp: 2100.00
                breakdown_by_service:
                  - service: "EC2"
                    cost: 5000.00
                  - service: "S3"
                    cost: 1200.00
                  - service: "RDS"
                    cost: 2000.00
                  - service: "Lambda"
                    cost: 800.00
        '400':
          description: Invalid date range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...